build: embed the client via a build tag and streamline the release pipeline
Build the UI straight into server/frontend/dist (Vite's outDir) and embed it
only in release binaries, gated by the `embedclient` build tag. Default builds
(CI test/lint/security) no longer need a built client nor a committed
placeholder: the stub reports no embedded UI and the server serves from
--static-files.

- go:embed moved behind //go:build embedclient; a stub covers default builds.
- Vite outputs to the embed dir directly (no build/client + cp); drop the
  Parcel-era assetsPrefix plugin — assetsDir "assets" + base "./" emits
  ./assets/<file> natively, served from dist/assets.
- Makefile: list release platforms in RELEASE_PLATFORMS and loop; build-docker
  builds only the native amd64 binary for the smoke test.
- CI: cross-compile release binaries only on tags (PRs build + smoke the native
  image); use github.ref_name instead of extract_ref; collapse artifact lists to
  a build/smocker-* glob; give lint setup-go.

The published multi-arch image (deploy-docker buildx) is unchanged.
