+++ server/frontend/frontend.go
@@ -1,27 +1,14 @@
-// Package frontend embeds the built web UI so a released binary is self-contained (no need to
-// ship the client directory alongside it). The dist directory is populated by the client build
-// (see the Makefile release target); a committed placeholder keeps `go build` working before the
-// client has ever been built.
+//go:build !embedclient
+
+// Package frontend optionally embeds the built web UI. By default (no build tag) the UI is NOT
+// embedded: FS reports none, so the server falls back to serving from --static-files. This lets
+// `go build` / `go test` work without a built client (CI test/lint/security never build it).
+// Release builds pass `-tags embedclient` after building the client to bake the UI into the
+// binary — see frontend_embed.go.
 package frontend
 
-import (
-	"embed"
-	"io/fs"
-)
+import "io/fs"
 
-//go:embed all:dist
-var dist embed.FS
-
-// FS returns the embedded client filesystem (rooted at dist) together with whether it actually
-// holds a built UI (an index.html). When only the placeholder is present it returns ok=false, so
-// callers fall back to serving from the --static-files directory on disk.
-func FS() (fs.FS, bool) {
-	sub, err := fs.Sub(dist, "dist")
-	if err != nil {
-		return nil, false
-	}
-	if _, err := fs.Stat(sub, "index.html"); err != nil {
-		return nil, false
-	}
-	return sub, true
-}
+// FS returns the embedded client filesystem and whether a UI is embedded. Without the
+// "embedclient" build tag there is none.
+func FS() (fs.FS, bool) { return nil, false }
