+++ vite.config.ts
@@ -1,35 +1,7 @@
 /// <reference types="vitest/config" />
-import { readFileSync, writeFileSync } from "node:fs";
-import { resolve } from "node:path";
 import react from "@vitejs/plugin-react";
 import type { Plugin } from "vite";
 import { defineConfig } from "vitest/config";
-
-// The Go admin server serves the built client directory (build/client) under the /assets URL
-// prefix and renders index.html as a Go text/template (injecting {{.basePath}} / {{.version}}).
-// Parcel used publicUrl "./assets": flat files in build/client referenced as relative
-// "./assets/<file>", so the runtime <base href> makes them resolve under any deployment base
-// path (e.g. behind Caddy stripping /smocker). Vite forbids a "./assets/" base, so we emit flat,
-// relative refs (base "./", assetsDir ".") and prefix them with "assets/" in the HTML afterwards.
-// Keep this in sync with server/admin_server.go (Static("/assets", ...) + renderIndex) and the
-// Makefile (which packages build/client).
-function assetsPrefix(): Plugin {
-  let indexPath = "";
-  return {
-    name: "smocker-assets-prefix",
-    apply: "build",
-    configResolved(cfg) {
-      indexPath = resolve(cfg.root, cfg.build.outDir, "index.html");
-    },
-    // Rewrite on disk after Vite has finished emitting (so both injected <script> tags and
-    // resolved source <link> hrefs are covered): "./<file>" -> "./assets/<file>".
-    closeBundle() {
-      let html = readFileSync(indexPath, "utf8");
-      html = html.replace(/(src|href)="\.\/(?!assets\/)/g, '$1="./assets/');
-      writeFileSync(indexPath, html);
-    },
-  };
-}
 
 // In the dev server, Vite serves index.html without the Go template step, so the runtime globals
 // stay as literal "{{.basePath}}" / "{{.version}}" and break the API base + <base href>. Substitute
@@ -57,10 +29,11 @@ const apiProxy = Object.fromEntries(
 
 export default defineConfig(({ command }) => ({
   root: "client",
-  // Dev serves from "/" (clean absolute URLs against the dev server); the build keeps "./" so the
-  // Go server's <base href> can relocate assets under any deployment base path.
+  // The build emits relative asset refs ("./assets/<file>") so the Go server's runtime
+  // <base href="{{.basePath}}"> can relocate them under any deployment base path (e.g. behind
+  // Caddy stripping /smocker). Dev serves from "/" for clean absolute URLs against the dev server.
   base: command === "build" ? "./" : "/",
-  plugins: [react(), assetsPrefix(), devIndexHtml()],
+  plugins: [react(), devIndexHtml()],
   server: {
     // Allow importing the canonical mock schema (docs/mock.schema.json) from client code, which
     // lives above the Vite root.
@@ -69,9 +42,12 @@ export default defineConfig(({ command }) => ({
     proxy: apiProxy,
   },
   build: {
-    outDir: "../build/client",
+    // Build straight into the go:embed source: release binaries bake it in (-tags embedclient),
+    // and dev serves it from disk via --static-files. Assets land in dist/assets (served at
+    // /assets/*), index.html at the root (rendered as a Go template). Keep in sync with
+    // server/admin_server.go and the Makefile.
+    outDir: "../server/frontend/dist",
     emptyOutDir: true,
-    assetsDir: ".",
   },
   test: {
     environment: "jsdom",
